Fix bugs in save
