+++ postmortem.go
@@ -1,12 +1,15 @@
 package postmortems
 
 import (
+	"bytes"
 	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
+	"log"
 	"os"
 	"path/filepath"
+	"text/template"
 	"time"
 
 	"github.com/gernest/front"
@@ -18,8 +21,8 @@ import (
 type Postmortem struct {
 	UUID        string    `yaml:"uuid"`
 	URL         string    `yaml:"url"`
-	StartTime   time.Time `yaml:"start_time"`
-	EndTime     time.Time `yaml:"end_time"`
+	StartTime   time.Time `yaml:"start_time,omitempty"`
+	EndTime     time.Time `yaml:"end_time,omitempty"`
 	Categories  []string  `yaml:"categories"`
 	Company     string    `yaml:"company"`
 	Product     string    `yaml:"product"`
@@ -147,7 +150,33 @@ func GenerateJSON(d string) error {
 	})
 }
 
+// ToYaml transforms a postmortem into yaml for the frontmatter.
 func ToYaml(pm *Postmortem) (string, error) {
 	bytes, err := yaml.Marshal(pm)
 	return string(bytes), err
+}
+
+// Save takes the in-memory representation of the postmortem file and stores it
+// in a file.
+func (pm *Postmortem) Save(dir string) error {
+	var data bytes.Buffer
+
+	fm := template.New("PostmortemTemplate")
+	fm = fm.Funcs(template.FuncMap{"yaml": ToYaml})
+	fm, err := fm.Parse(bodyTmpl)
+	if err != nil {
+		return err
+	}
+
+	if err := fm.Execute(&data, pm); err != nil {
+		return fmt.Errorf("error executing template: %w", err)
+	}
+
+	// Write postmortem data from memory to file.
+	if err := ioutil.WriteFile(filepath.Join(dir, pm.UUID+".md"), data.Bytes(), 0644); err != nil {
+		return fmt.Errorf("error writing file: %w", err)
+	}
+	log.Printf("saved %+v to %+v", pm, data)
+
+	return nil
 }
